lexers/internal: remove danwakefield/fnmatch
This package appears to be old and unmaintained, and it gets pulled into the
go.sum of every single package importing Chroma. Also fnmatch doesnt seem to do
any error checking on the pattern, while the standard package does. Also, the
standard package path/filepath is already important, so it appears not to cost
anything to change it.