fix(dns): highlight domain names as NameClass
This commit fixes #885.
Before this commit, only the first part (before the first dot) of a
domain name was highlighted as NameClass``, which made the rest
highlight in `err` (red). This commit makes domain names anywhere in the
line highlight as NameClass, still allowing @ at the start of a line to
be highlighted just as before.