Disable Default Clear background for TTY
I want to be able to have a background colour for the output in tty.
This would help in getting a nicer looking code snippet in projects
like https://charm.sh/.

The functionality of clearing the background is still present. But
now it is done explicitly though the style.ClearBackgound() function.