fix: Dot in a string breaks the highlighting in Terraform. Fixes #988 (#1065)
Admittedly, I set how to make Chroma look more like Pygments.

Pygments:


![pygments-string-capture](https://github.com/user-attachments/assets/8f19bfc6-533c-41d8-859b-219fc13d506e)

Chroma:


![chroma-string-current-capture](https://github.com/user-attachments/assets/eda1fd37-d0e1-4b69-a92c-03299d540663)

Pygments in my opinion does this correctly, with everything between `"
... "` is a string. Chroma does some inner interpolation to pull out
those inner pieces. I can see both sides and in that spirit, I chose to
keep the current convention. I do have a branch that will make Chroma
like Pygments if that's the desired behavior.

While looking into existing Terraform issues and Pull requests, I
encountered #988 and wondered if I could solve it.

As it currently exists:


![988-current-capture](https://github.com/user-attachments/assets/ff78195d-c9b6-4384-8395-f0de9c10bd97)

Fixed in this PR:


![chroma-988-fixed-capture](https://github.com/user-attachments/assets/fd8db050-f3e1-4804-8717-f29f6e411e17)

A silver lining is punctuation actually works and mirrors Pygments.
There is a possible regression depending on one's point of view and I
wanted to call it out or not hide it.

Current:


![current-chroma-iam-policy-doc-github-dark-capture](https://github.com/user-attachments/assets/707f8c78-f440-44cb-beea-b29b3c913d5c)

This PR:


![chroma-proposed-fixed-iam-policy-document-github-dark-capture](https://github.com/user-attachments/assets/c7927eb9-a4a0-4a28-a399-371b6c49c807)

The punctuation inside strings is highlighted differently. I don't know
whether this solution is worth it or if it is a regression not worth
considering, of if it is even a regression at all.

But I've spent too much time and admittedly, I don't understand the
syntax well enough. I solved my problem, maybe another one, maybe not.
I'll understand if you close this PR, perhaps someone else can pick it
up and address the regression if it is considered a regression. If you
treat everything inside `" ... "` as a string, then this possible
regression becomes moot.