fix: properly tokenizes default_tags attribute (#1068)
In the current version `default_tags` was getting processed like

```json
{"type":"NameAttribute","value":"default"},
{"type":"Error","value":"_"},
{"type":"NameOther","value":"tags"}
```

but `default_tags` is a valid attribute or parameter here and doesn't
need parsed as such. This matches `default_tags` properly

Here is the before

![default_tags_2_16_capture](https://github.com/user-attachments/assets/fccf327a-82d5-4cc0-842e-acc2a605dce7)

And the after with this change

![fixed-default_tags-capture](https://github.com/user-attachments/assets/8c8f11cb-4e46-469e-bdce-6ddfc9a2d600)

Hopefully I did this correctly. I fixed #988 but it's been a minute.
Thanks!