fix(vue): handle more edge cases for tags (#1069)
- Allow parsing of attributes to end on `/>`, `>` or a whitespace
character.
- Consilidate the parsing of the shorthand `@EVENT` to one rule. Do
better parsing for its contents by allowing whitespace
(`@click="menuVisible = !menuVisible"`) and `EVENT` to contain a dot
(`@keyup.enter`). Allow it to not have any condition (`<a
@click.stop>`).
- Allow `v-` directives to contain a dash.
- Move a rule down that captures `v-on:click="doSomething"` as it was
included more than it should for the `LiteralString` token.
- Add tests that should exhaustively tests these edge cases. They were
derived from
https://codeberg.org/forgejo/forgejo/src/commit/a18fc8b03702f439845daf22a0cc7fb0f8ea95c3/web_src/js/components/RepoBranchTagSelector.vue#L248-L318

## Screenshot
The syntax highlighting of the input is still not perfect (e.g. HTML
comment and `symbol-id=`), but it's a lot better now.

### Before

![image](https://github.com/user-attachments/assets/fc297510-ccfb-4a64-90f6-768afa17459e)

### After

![image](https://github.com/user-attachments/assets/7d9f2adc-263f-4dd1-938d-927de819369c)