Added IBM RPG lexer (#1078)
Hi !

Here's a lexer for IBM RPG !
You can learn about this language here :
- https://en.wikipedia.org/wiki/IBM_RPG
- https://www.ibm.com/docs/en/i/7.6.0?topic=languages-rpg

For legacy purposes, the language can be written in two different forms,
the fixed format (looks like a punched card, expects specific characters
to be at certain columns in source code) and the free format (gets rid
of fixed format, and looks like modern languages as Python).
This lexer only handles the free format, I may open a pull request later
to implement the fixed format syntax.

There are some RPG repositories available on GitHub, you can search
using "language:RPGLE" to see them.

The language is old but still maintained and even continues to evolve
nowadays.
The lexer is based on the latest version at this day, RPG 7.6.

Are the MIME types in the XML files important ? I don't think this
language has one (and I didn't know there were MIME types for
programming languages), so I made one which looks like what I saw in
other lexers XML files.

And the language can have embedded SQL statements in the source code, so
I reused the SQL lexer. Please tell me if there's a better way to do it.

The language is rather tied to IBM environment, so if you're interested,
here's how to try it :
- https://pub400.com/ provides free accounts to an IBM machine to try
and discover it
-
https://www.nicklitten.com/use-a-5250-green-screen-terminal-in-visual-studio/
to get a terminal to access pub400 machine and run code
-
https://www.nicklitten.com/connecting-to-pub400-with-vs-code-for-ibm-i/
shows how to code RPG in VSCode and compile it