fix: don't emit empty tail tokens (#1121)
When tokens are split into lines, tokens that end with a newline are
emitted again as an empty token, which is confusing and doesn't have any
benefit. This conversion shouldn't emit a empty tail token.

Adds a test.

I noticed this issue by a CI failure for the output of a Go program that
was changed because of
d0ad67944442bcb84ff4710fe8dfa8f76ecedc01 and the new output contained a
empty whitespace token at the beginning of most lines.