more idiomatic iterator use
nerdsnipe alert: I'm still in relatively early phases of exploring iterator
idiom, so couldn't resist having a go at simplifying SplitTokensIntoLines
which is now fully streaming (even though none of the call sites
look like they can take advantage of that currently). Incidentally, it's
now O(n) not O(n^2) too (that `for strings.Contains` looked expensive).

I removed the `EOF` variable, replacing it with an `IsZero` method
which can be used to check if a `Token` is valid or not.

I don't see anywhere that seems to rely on yielding an EOF value:
all the tests pass.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>