Add helper script to generate supported langs markdown table (#1170)
**TL;DR: added a python script that generates a "supported languages"
markdown table, making it easier to keep the table in the readme up to
date**

---

This PR adds a Python script `_tools/format_supported_langs.py` which
accepts the output of `chroma --list` over stdin and outputs a Markdown
table which can be added to the Supported Languages section of the
README. Here's how I called the script:

```sh
env -C cmd/chroma go run . | uv run _tools/format_supported_langs.py
```

I chose not to try automating edits to the README, and leave it up to
the invoker to replace the table in the readme with this script's
output. I also updated the table in the README because it's missing a
few languages.

Hopefully this can help with the maintenance burden.