Update bash.xml
Change the name of the language to "Shell".

Bash, while extremely widespread, is a richer language than any shell. Having the lexer report a shell script as a Bash script can be confusing, or even error-inducing.

For correctness, the only thing the lexer can (correctly) assume is that the file is shell script.

Beyond that, a syntactic analysis would be required.

I guess the file name should be changed too.